# api.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: User Service
  license:
    name: MIT
servers:
  - url: http://localhost
paths:
  /register:
    post:
      summary: Register a new user
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserRequest"
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterUserResponse"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /login:
    post:
      summary: Login with phone number and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginUserRequest"
      responses:
        '200':
          description: User logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginUserResponse"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /profile:
    get:
      summary: Get user profile
      operationId: getUserProfile
      security:
        - Authorization: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetUserProfileResponse"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      summary: Update user profile
      operationId: updateUserProfile
      security:
        - Authorization: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserProfileRequest"
      responses:
        '200':
          description: User profile updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateUserProfileResponse"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    Authorization:
      type: http
      scheme: bearer
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        error:
          type: object
    RegisterUserRequest:
      type: object
      properties:
        phone_number:
          type: string
        full_name:
          type: string
        password:
          type: string
      required:
        - phone_number
        - full_name
        - password
    RegisterUserResponse:
      type: object
      properties:
        id:
          type: integer
    LoginUserRequest:
      type: object
      properties:
        phone_number:
          type: string
        password:
          type: string
      required:
        - phone_number
        - password
    LoginUserResponse:
      type: object
      properties:
        id:
          type: integer
        token:
          type: string
    GetUserProfileResponse:
      type: object
      properties:
        phone_number:
          type: string
        full_name:
          type: string
    UpdateUserProfileRequest:
      type: object
      properties:
        phone_number:
          type: string
        full_name:
          type: string
    UpdateUserProfileResponse:
      type: object
      properties:
        phone_number:
          type: string
        full_name:
          type: string